Better service names; reporter and listener classes marked as final
