+++ src/Resources/config/services.yml
@@ -2,10 +2,10 @@ services:
     _defaults:
         autowire: true
 
-    Crassula\Bundle\GoogleCloudErrorReportingBundle\Service\ErrorNotifier:
-        '@Crassula\Bundle\GoogleCloudErrorReportingBundle\Service\ErrorReporter'
-
     Crassula\Bundle\GoogleCloudErrorReportingBundle\Service\ErrorReporter:
+        '@Crassula\Bundle\GoogleCloudErrorReportingBundle\Service\GoogleCloudErrorReporter'
+
+    Crassula\Bundle\GoogleCloudErrorReportingBundle\Service\GoogleCloudErrorReporter:
         public: true
         arguments:
             $config: '%crassula_google_cloud_error_reporting.config%'
