--- src/Service/ErrorNotifier.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Crassula\Bundle\GoogleCloudErrorReportingBundle\Service;
-
-interface ErrorNotifier
-{
-    public function report(\Throwable $error, array $options = []): bool;
-}
